relocate pattern_engines.js
